+++ packages/gcp/src/__tests__/ServiceWrapper.test.ts
@@ -3,33 +3,36 @@
  */
 
 import { ProjectsClient } from '@google-cloud/resource-manager'
-import { compute_v1 } from 'googleapis'
-import { compute as googleCompute } from 'googleapis/build/src/apis/compute'
+import {
+  AddressesClient,
+  DisksClient,
+  ImagesClient,
+  InstancesClient,
+  MachineTypesClient,
+  protos as googleCompute,
+} from '@google-cloud/compute'
 import { GoogleAuth } from 'google-auth-library'
 import { RecommenderClient } from '@google-cloud/recommender'
 import { GoogleAuthClient, wait } from '@cloud-carbon-footprint/common'
 import {
-  InstanceData,
-  mockedAddressesResultItems,
+  mockAddressesResultItems,
   mockedAddressGetDetails,
   mockedDisksGetSSDDetails,
-  mockedDisksResultItems,
+  mockDisksResultItems,
   mockedImageGetDetails,
   mockedInstanceGetItems,
-  mockedInstanceResultItems,
+  mockInstanceResultItems,
   mockedMachineTypesGetItems,
 } from './fixtures/googleapis.fixtures'
-import Schema$Instance = compute_v1.Schema$Instance
-import Schema$MachineType = compute_v1.Schema$MachineType
-
+import { mockedProjects } from './fixtures/resourceManager.fixtures'
+import { mockStopVMRecommendationsResults } from './fixtures/recommender.fixtures'
+import Instance = googleCompute.google.cloud.compute.v1.IInstance
+import MachineType = googleCompute.google.cloud.compute.v1.IMachineType
 import ServiceWrapper from '../lib/ServiceWrapper'
 import {
   ActiveProject,
   RecommenderRecommendations,
 } from '../lib/RecommendationsTypes'
-
-import { mockStopVMRecommendationsResults } from './fixtures/recommender.fixtures'
-import { mockedProjects } from './fixtures/resourceManager.fixtures'
 import { setupSpy, setupSpyWithRejectedValue } from './helpers'
 
 jest.mock('@cloud-carbon-footprint/common', () => ({
@@ -67,39 +70,38 @@ describe('GCP Service Wrapper', () => {
     ;(getClientSpy as jest.Mock).mockResolvedValue(jest.fn())
 
     const googleAuthClient: GoogleAuthClient = await auth.getClient()
-    const googleComputeClient = googleCompute('v1')
 
     serviceWrapper = new ServiceWrapper(
       new ProjectsClient(),
       googleAuthClient,
-      googleComputeClient,
+      new InstancesClient(),
+      new DisksClient(),
+      new AddressesClient(),
+      new ImagesClient(),
+      new MachineTypesClient(),
       new RecommenderClient(),
     )
 
     setupSpy(
-      googleComputeClient.instances,
-      'aggregatedList',
-      mockedInstanceResultItems,
+      InstancesClient.prototype,
+      'aggregatedListAsync',
+      mockInstanceResultItems(),
     )
     setupSpy(
-      googleComputeClient.disks,
-      'aggregatedList',
-      mockedDisksResultItems,
+      DisksClient.prototype,
+      'aggregatedListAsync',
+      mockDisksResultItems(),
     )
-    setupSpy(googleComputeClient.disks, 'get', mockedDisksGetSSDDetails)
+    setupSpy(DisksClient.prototype, 'get', mockedDisksGetSSDDetails)
     setupSpy(
-      googleComputeClient.addresses,
-      'aggregatedList',
-      mockedAddressesResultItems,
+      AddressesClient.prototype,
+      'aggregatedListAsync',
+      mockAddressesResultItems(),
     )
-    setupSpy(
-      googleComputeClient.machineTypes,
-      'get',
-      mockedMachineTypesGetItems,
-    )
-    setupSpy(googleComputeClient.instances, 'get', mockedInstanceGetItems)
-    setupSpy(googleComputeClient.images, 'get', mockedImageGetDetails)
-    setupSpy(googleComputeClient.addresses, 'get', mockedAddressGetDetails)
+    setupSpy(MachineTypesClient.prototype, 'get', mockedMachineTypesGetItems)
+    setupSpy(InstancesClient.prototype, 'get', mockedInstanceGetItems)
+    setupSpy(ImagesClient.prototype, 'get', mockedImageGetDetails)
+    setupSpy(AddressesClient.prototype, 'get', mockedAddressGetDetails)
   })
 
   it('gets active projects', async () => {
@@ -148,28 +150,25 @@ describe('GCP Service Wrapper', () => {
   })
 
   it('gets instance details', async () => {
-    const instanceDetails: Schema$Instance =
-      await serviceWrapper.getInstanceDetails(
-        'project',
-        'test-instance',
-        'us-west1-b',
-      )
+    const instanceDetails: Instance = await serviceWrapper.getInstanceDetails(
+      'project',
+      'test-instance',
+      'us-west1-b',
+    )
 
-    const expectedResult: InstanceData = {
-      data: {
-        machineType:
-          'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/n2-standard-32',
-        disks: [],
-        id: '12456789012',
-        name: 'test-resource-name',
-      },
+    const expectedResult: Instance = {
+      machineType:
+        'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/n2-standard-32',
+      disks: [],
+      id: '12456789012',
+      name: 'test-resource-name',
     }
 
-    expect(instanceDetails).toEqual(expectedResult.data)
+    expect(instanceDetails).toEqual(expectedResult)
   })
 
   it('gets machine type details', async () => {
-    const machineTypeDetails: Schema$MachineType =
+    const machineTypeDetails: MachineType =
       await serviceWrapper.getMachineTypeDetails(
         'project',
         'test-instance',
@@ -241,7 +240,6 @@ describe('GCP Service Wrapper', () => {
 
   describe('error handling', () => {
     let serviceWrapper: ServiceWrapper
-    const googleComputeClient = googleCompute('v1')
 
     beforeEach(async () => {
       const auth = new GoogleAuth({
@@ -257,15 +255,19 @@ describe('GCP Service Wrapper', () => {
       serviceWrapper = new ServiceWrapper(
         new ProjectsClient(),
         googleAuthClient,
-        googleComputeClient,
+        new InstancesClient(),
+        new DisksClient(),
+        new AddressesClient(),
+        new ImagesClient(),
+        new MachineTypesClient(),
         new RecommenderClient(),
       )
     })
 
     it('fails to get active zones for project', async () => {
       setupSpyWithRejectedValue(
-        googleComputeClient.instances,
-        'aggregatedList',
+        InstancesClient.prototype,
+        'aggregatedListAsync',
         'error',
       )
       const activeProjectsAndZones: ActiveProject[] =