+++ packages/gcp/src/__tests__/fixtures/googleapis.fixtures.ts
@@ -1,130 +1,113 @@
 /*
  * © 2021 Thoughtworks, Inc.
  */
-import { compute_v1 } from 'googleapis'
-import Schema$Instance = compute_v1.Schema$Instance
-import Schema$MachineType = compute_v1.Schema$MachineType
-import Schema$Image = compute_v1.Schema$Image
-import Schema$InstanceAggregatedList = compute_v1.Schema$InstanceAggregatedList
-import Schema$Disk = compute_v1.Schema$Disk
-import Schema$DiskAggregatedList = compute_v1.Schema$DiskAggregatedList
-import Schema$AddressAggregatedList = compute_v1.Schema$AddressAggregatedList
-import Schema$Address = compute_v1.Schema$Address
 
-export type InstanceData = {
-  data: Schema$Instance
+import { ScopedListResult } from '../../lib/ServiceWrapperTypes'
+import { google } from '@google-cloud/compute/build/protos/protos'
+import IDisk = google.cloud.compute.v1.IDisk
+import IInstance = google.cloud.compute.v1.IInstance
+import IMachineType = google.cloud.compute.v1.IMachineType
+import IImage = google.cloud.compute.v1.IImage
+import IAddress = google.cloud.compute.v1.IAddress
+
+interface IterableMockResponse {
+  next(): Promise<IteratorResult<ScopedListResult>>
 }
 
-type MachineTypeData = {
-  data: Schema$MachineType
+const mockAsyncIterator = {
+  // * = fancy generator function (https://javascript.info/generators)
+  async *[Symbol.asyncIterator](mockResponse: ScopedListResult[]) {
+    for (const response of mockResponse) yield response
+  },
 }
 
-type ImageDetails = {
-  data: Schema$Image
+export const mockIterableResponse = (
+  mockResponse: ScopedListResult[],
+): IterableMockResponse => {
+  return mockAsyncIterator[Symbol.asyncIterator](mockResponse)
 }
 
-type DiskData = {
-  data: Schema$Disk
-}
-
-type AddressDetails = {
-  data: Schema$Address
-}
-
-type InstanceAggregatedList = {
-  data: Schema$InstanceAggregatedList
-}
-
-type DiskAggregatedList = {
-  data: Schema$DiskAggregatedList
-}
-
-type AddressAggregatedList = {
-  data: Schema$AddressAggregatedList
-}
-
-export const mockedInstanceResultItems: InstanceAggregatedList = {
-  data: {
-    items: {
-      'zones/us-west1-a': {
+export const mockInstanceResultItems = () =>
+  mockIterableResponse([
+    [
+      'zones/us-west1-a',
+      {
         instances: [{ id: 'test-instance' }],
       },
-      'zones/us-east1-a': {
+    ],
+    [
+      'zones/us-east1-a',
+      {
         instances: [{ id: 'test-instance-1' }],
       },
-      'zones/us-west1-b': { warning: { code: 'NO_RESULTS_ON_PAGE' } },
-    },
-  },
-}
+    ],
+    ['zones/us-west1-b', { warning: { code: 'NO_RESULTS_ON_PAGE' } }],
+  ])
 
-export const mockedInstanceRegionsResultItems: InstanceAggregatedList = {
-  data: {
-    items: {
-      'regions/us-west1': {
+export const mockInstanceRegionsResultItems = () =>
+  mockIterableResponse([
+    [
+      'regions/us-west1',
+      {
         instances: [{ id: 'test-instance' }],
       },
-    },
-  },
-}
+    ],
+  ])
 
-export const mockedInstanceGlobalResultItems: InstanceAggregatedList = {
-  data: {
-    items: {
-      global: {
+export const mockInstanceGlobalResultItems = () =>
+  mockIterableResponse([
+    [
+      'global',
+      {
         instances: [{ id: 'test-instance-global' }],
       },
-    },
-  },
-}
+    ],
+  ])
 
-export const mockedAddressesResultItems: AddressAggregatedList = {
-  data: { items: {} },
-}
+export const mockAddressesResultItems = () => mockIterableResponse([])
 
-export const mockedDisksResultItems: DiskAggregatedList = {
-  data: { items: {} },
-}
+export const mockDisksResultItems = () => mockIterableResponse([])
 
-export const mockedDisksGetSSDDetails: DiskData = {
-  data: {
+export const mockedDisksGetSSDDetails: [IDisk] = [
+  {
     sizeGb: '20',
     type: 'https://www.googleapis.com/compute/v1/projects/techops-events/zones/us-central1-b/diskTypes/pd-standard-ssd',
     id: '12456789012',
     name: 'test-resource-name',
   },
-}
+]
 
-export const mockedDisksGetHDDDetails: DiskData = {
-  data: {
+export const mockedDisksGetHDDDetails: [IDisk] = [
+  {
     sizeGb: '20',
     type: 'https://www.googleapis.com/compute/v1/projects/techops-events/zones/us-central1-b/diskTypes/pd-standard',
     id: '12456789012',
     name: 'test-resource-name',
   },
-}
+]
 
-export const mockedInstanceGetItems: InstanceData = {
-  data: {
+export const mockedInstanceGetItems: [IInstance] = [
+  {
     machineType:
       'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/n2-standard-32',
     disks: [],
     id: '12456789012',
     name: 'test-resource-name',
   },
-}
+]
 
-export const mockedInstanceGetItemsCurrent: InstanceData = {
-  data: {
+export const mockedInstanceGetItemsCurrent: [IInstance] = [
+  {
     machineType:
       'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/e2-medium',
     disks: [],
     id: '12456789012',
     name: 'test-resource-name',
   },
-}
+]
 
-export const mockedInstanceGetItemsWithHDDDisks: InstanceData = {
-  data: {
+export const mockedInstanceGetItemsWithHDDDisks: [IInstance] = [
+  {
     machineType:
       'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/n2-standard-32',
     disks: [
@@ -137,10 +120,10 @@ export const mockedInstanceGetItemsWithHDDDisks: InstanceData = {
     id: '12456789012',
     name: 'test-instance-name',
   },
-}
+]
 
-export const mockedInstanceGetItemsWithBothDisks: InstanceData = {
-  data: {
+export const mockedInstanceGetItemsWithBothDisks: [IInstance] = [
+  {
     machineType:
       'https://www.googleapis.com/compute/v1/projects/test-project/zones/us-west1-b/machineTypes/n2-standard-32',
     disks: [
@@ -158,38 +141,38 @@ export const mockedInstanceGetItemsWithBothDisks: InstanceData = {
     id: '12456789012',
     name: 'test-instance-name',
   },
-}
+]
 
-export const mockedMachineTypesGetItems: MachineTypeData = {
-  data: {
+export const mockedMachineTypesGetItems: [IMachineType] = [
+  {
     guestCpus: 32,
   },
-}
+]
 
-export const mockedMachineTypesGetItemsCurrent: MachineTypeData = {
-  data: {
+export const mockedMachineTypesGetItemsCurrent: [IMachineType] = [
+  {
     guestCpus: 2,
   },
-}
+]
 
-export const mockedMachineTypesGetItemsNew: MachineTypeData = {
-  data: {
+export const mockedMachineTypesGetItemsNew: [IMachineType] = [
+  {
     guestCpus: 1,
   },
-}
+]
 
-export const mockedImageGetDetails: ImageDetails = {
-  data: {
+export const mockedImageGetDetails: [IImage] = [
+  {
     archiveSizeBytes: '580709696',
     id: '12456789012',
     name: 'test-resource-name',
   },
-}
+]
 
-export const mockedAddressGetDetails: AddressDetails = {
-  data: {
+export const mockedAddressGetDetails: [IAddress] = [
+  {
     id: '123456789012345',
     name: 'test-address',
     address: '38.141.210.105',
   },
-}
+]