+++ packages/gcp/src/lib/Recommendations.ts
@@ -3,12 +3,12 @@
  */
 
 import R from 'ramda'
-import { compute_v1 } from 'googleapis'
-import { google } from '@google-cloud/recommender/build/protos/protos'
-import IRecommendation = google.cloud.recommender.v1.IRecommendation
-import IImpact = google.cloud.recommender.v1.IImpact
-import Schema$Instance = compute_v1.Schema$Instance
-import Schema$Disk = compute_v1.Schema$Disk
+import { google as googleRecommender } from '@google-cloud/recommender/build/protos/protos'
+import { protos as googleCompute } from '@google-cloud/compute'
+import IRecommendation = googleRecommender.cloud.recommender.v1.IRecommendation
+import IImpact = googleRecommender.cloud.recommender.v1.IImpact
+import Instance = googleCompute.google.cloud.compute.v1.IInstance
+import Disk = googleCompute.google.cloud.compute.v1.IDisk
 import {
   COMPUTE_PROCESSOR_TYPES,
   ComputeEstimator,
@@ -16,6 +16,7 @@ import {
   ICloudRecommendationsService,
   StorageEstimator,
   StorageUsage,
+  KilowattHourTotals,
 } from '@cloud-carbon-footprint/core'
 import {
   convertBytesToGigabytes,
@@ -36,7 +37,6 @@ import {
 } from './RecommendationsTypes'
 import ServiceWrapper from './ServiceWrapper'
 import { GCP_REGIONS } from './GCPRegions'
-import { KilowattHourTotals } from '@cloud-carbon-footprint/core'
 
 export default class Recommendations implements ICloudRecommendationsService {
   readonly RECOMMENDER_IDS: string[] = [
@@ -213,7 +213,7 @@ export default class Recommendations implements ICloudRecommendationsService {
           .split('/')
           .pop(),
     }
-    let diskDetails: Schema$Disk
+    let diskDetails: Disk
     try {
       switch (recommendation.recommenderSubtype) {
         case RECOMMENDATION_TYPES.STOP_VM:
@@ -253,7 +253,7 @@ export default class Recommendations implements ICloudRecommendationsService {
             timestamp: undefined,
           }
           resourceDetails = {
-            resourceId: instanceDetails.id,
+            resourceId: instanceDetails.id.toString(),
             resourceName: instanceDetails.name,
           }
           return [footprintEstimate, resourceDetails]
@@ -291,7 +291,7 @@ export default class Recommendations implements ICloudRecommendationsService {
             zone,
           )
           resourceDetails = {
-            resourceId: currentMachineInstanceDetails.id,
+            resourceId: currentMachineInstanceDetails.id.toString(),
             resourceName: currentMachineInstanceDetails.name,
           }
           return [footprintEstimate, resourceDetails]
@@ -309,7 +309,7 @@ export default class Recommendations implements ICloudRecommendationsService {
           )
 
           resourceDetails = {
-            resourceId: diskDetails.id,
+            resourceId: diskDetails.id.toString(),
             resourceName: diskDetails.name,
           }
           return [footprintEstimate, resourceDetails]
@@ -320,14 +320,14 @@ export default class Recommendations implements ICloudRecommendationsService {
             imageId,
           )
           const imageArchiveSizeGigabytes = convertBytesToGigabytes(
-            parseFloat(imageDetails.archiveSizeBytes),
+            parseFloat(imageDetails.archiveSizeBytes.toString()),
           )
           footprintEstimate = this.estimateStorageCO2eSavings(
             imageArchiveSizeGigabytes,
             this.parseRegionFromZone(zone),
           )
           resourceDetails = {
-            resourceId: imageDetails.id,
+            resourceId: imageDetails.id.toString(),
             resourceName: imageDetails.name,
           }
           return [footprintEstimate, resourceDetails]
@@ -340,7 +340,7 @@ export default class Recommendations implements ICloudRecommendationsService {
               zone,
             )
           resourceDetails = {
-            resourceId: addressDetails.id,
+            resourceId: addressDetails.id.toString(),
             resourceName: addressDetails.name,
           }
 
@@ -383,15 +383,12 @@ export default class Recommendations implements ICloudRecommendationsService {
     )
   }
 
-  private async getCO2EstimatesSavingsForDisk(
-    diskDetails: Schema$Disk,
-    zone: string,
-  ) {
+  private async getCO2EstimatesSavingsForDisk(diskDetails: Disk, zone: string) {
     const storageType = this.googleServiceWrapper.getStorageTypeFromDiskName(
       diskDetails.type.split('/').pop(),
     )
     return this.estimateStorageCO2eSavings(
-      parseFloat(diskDetails.sizeGb),
+      parseFloat(diskDetails.sizeGb.toString()),
       this.parseRegionFromZone(zone),
       storageType,
     )
@@ -399,7 +396,7 @@ export default class Recommendations implements ICloudRecommendationsService {
 
   private async getCO2EstimatedSavingsForInstance(
     projectId: string,
-    instanceDetails: Schema$Instance,
+    instanceDetails: Instance,
     zone: string,
   ) {
     const machineType = instanceDetails.machineType.split('/').pop()