+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cloud-carbon-footprint/cli
 
+## 1.10.2
+
+### Patch Changes
+
+- Updated dependencies [0d20b32c]
+  - @cloud-carbon-footprint/common@1.11.0
+  - @cloud-carbon-footprint/app@1.2.2
+
+## 1.10.1
+
+### Patch Changes
+
+- 65386330: bumps ramda dependency
+- Updated dependencies [65386330]
+  - @cloud-carbon-footprint/app@1.2.1
+
 ## 1.10.0
 
 ### Minor Changes